requirements.txt
fastapi==0.128.8
psycopg2-binary==2.9.11
pydantic==2.12.5
SQLAlchemy==2.0.46
uvicorn==0.39.0
clickhouse-connect==0.11.0
python-dotenv==1.0.1
google-generativeai